First commit for rebasing the training/IOU CorDapp to M10
