--- .idea/runConfigurations/Run_Template_RPC_Client___Java.xml
@@ -1,17 +1,0 @@
-<component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="Run Template RPC Client - Java" type="Application" factoryName="Application" singleton="true">
-    <extension name="coverage" enabled="false" merge="false" sample_coverage="true" runner="idea" />
-    <option name="MAIN_CLASS_NAME" value="com.template.client.TemplateClientRPC" />
-    <option name="VM_PARAMETERS" value="" />
-    <option name="PROGRAM_PARAMETERS" value="localhost:10006" />
-    <option name="WORKING_DIRECTORY" value="" />
-    <option name="ALTERNATIVE_JRE_PATH_ENABLED" value="false" />
-    <option name="ALTERNATIVE_JRE_PATH" />
-    <option name="ENABLE_SWING_INSPECTOR" value="false" />
-    <option name="ENV_VARIABLES" />
-    <option name="PASS_PARENT_ENVS" value="true" />
-    <module name="java-source_main" />
-    <envs />
-    <method />
-  </configuration>
-</component>
