--- .idea/runConfigurations/Run_Template_RPC_Client___Kotlin.xml
@@ -1,15 +1,0 @@
-<component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="Run Template RPC Client - Kotlin" type="JetRunConfigurationType" factoryName="Kotlin" singleton="true">
-    <extension name="coverage" enabled="false" merge="false" sample_coverage="true" runner="idea" />
-    <option name="MAIN_CLASS_NAME" value="com.template.client.TemplateClientRPCKt" />
-    <option name="VM_PARAMETERS" value="" />
-    <option name="PROGRAM_PARAMETERS" value="localhost:10006" />
-    <option name="WORKING_DIRECTORY" value="" />
-    <option name="ALTERNATIVE_JRE_PATH_ENABLED" value="false" />
-    <option name="ALTERNATIVE_JRE_PATH" />
-    <option name="PASS_PARENT_ENVS" value="true" />
-    <module name="kotlin-source_main" />
-    <envs />
-    <method />
-  </configuration>
-</component>
