--- java-source/build.gradle
@@ -1,132 +1,0 @@
-apply plugin: 'java'
-apply plugin: 'kotlin'
-apply plugin: 'idea'
-apply plugin: 'net.corda.plugins.publish-utils'
-apply plugin: 'net.corda.plugins.cordformation'
-apply plugin: 'maven-publish'
-
-repositories {
-    mavenLocal()
-    mavenCentral()
-    jcenter()
-    maven { url 'https://dl.bintray.com/kotlin/exposed' }
-    maven { url 'https://jitpack.io' }
-}
-
-sourceSets {
-    main {
-        resources {
-            srcDir "../config/dev"
-        }
-    }
-    test {
-        resources {
-            srcDir "../config/test"
-        }
-    }
-}
-
-dependencies {
-    compile "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-    testCompile "junit:junit:$junit_version"
-
-    // Corda integration dependencies
-    compile "net.corda:core:$corda_version"
-    compile "net.corda:finance:$corda_version"
-    compile "net.corda:jackson:$corda_version"
-    compile "net.corda:node:$corda_version"
-    compile "net.corda:rpc:$corda_version"
-    compile "net.corda:webserver:$corda_version"
-    runtime "net.corda:corda:$corda_version"
-    runtime "net.corda:corda-webserver:$corda_version"
-
-    testCompile "net.corda:test-utils:$corda_version"
-
-    // GraphStream: For visualisation (required by TemplateClientRPC app)
-    compile "org.graphstream:gs-core:1.3"
-    compile("org.graphstream:gs-ui:1.3") {
-        exclude group: "bouncycastle"
-    }
-
-    // CorDapp dependencies
-    // Specify your cordapp's dependencies below, including dependent cordapps
-}
-
-task deployNodes(type: net.corda.plugins.Cordform, dependsOn: ['build']) {
-    directory "./build/nodes"
-    networkMap "Controller"
-    node {
-        name "Controller"
-        nearestCity "London"
-        advertisedServices = ["corda.notary.validating"]
-        p2pPort 10002
-        rpcPort 10003
-        webPort 10004
-        cordapps = []
-    }
-    node {
-        name "NodeA"
-        nearestCity "London"
-        advertisedServices = []
-        p2pPort 10005
-        rpcPort 10006
-        webPort 10007
-        cordapps = []
-        rpcUsers = [[ user: "user1", "password": "test", "permissions": []]]
-    }
-    node {
-        name "NodeB"
-        nearestCity "New York"
-        advertisedServices = []
-        p2pPort 10008
-        rpcPort 10009
-        webPort 10010
-        cordapps = []
-        rpcUsers = [[ user: "user1", "password": "test", "permissions": []]]
-    }
-    node {
-        name "NodeC"
-        nearestCity "San Francisco"
-        advertisedServices = []
-        p2pPort 10011
-        rpcPort 10012
-        webPort 10013
-        cordapps = []
-        rpcUsers = [[ user: "user1", "password": "test", "permissions": []]]
-    }
-}
-
-// TODO: Make into gradle plugin without any references to Jython
-task installJythonDeps(dependsOn: ['build']) {
-    project.copy {
-        from project.configurations.runtime
-        into "build/jythonDeps"
-    }
-}
-
-installJythonDeps.shouldRunAfter build
-
-idea {
-    module {
-        downloadJavadoc = true // defaults to false
-        downloadSources = true
-    }
-}
-
-publishing {
-    publications {
-        jarAndSources(MavenPublication) {
-            from components.java
-            artifactId 'cordapp-name-goes-here'
-
-            artifact sourceJar
-            artifact javadocJar
-        }
-    }
-}
-
-task runTemplateClientRPC(type: JavaExec) {
-    classpath = sourceSets.main.runtimeClasspath
-    main = 'com.template.client.TemplateClientRPC'
-    args 'localhost:10006'
-}
