--- java-source/src/main/resources/META-INF/services/net.corda.core.node.CordaPluginRegistry
@@ -1,2 +1,0 @@
-# Register a ServiceLoader service extending from net.corda.node.CordaPluginRegistry
-com.template.plugin.TemplatePlugin
