--- java-source/src/main/resources/certificates/readme.txt
@@ -1,1 +1,0 @@
-These certificates are used for development mode only (and are copies of those contained within the TraderDemo jar file)
