--- java-source/src/main/resources/templateWeb/index.html
@@ -1,25 +1,0 @@
-<!--
---------
-WARNING:
---------
-FOR SECURITY REASONS, USING A JAVASCRIPT WEB APP HOSTED VIA THE CORDA NODE IS NOT THE RECOMMENDED WAY TO INTERFACE
-WITH CORDA NODES! HOWEVER, FOR THIS PRE-ALPHA RELEASE IT'S A USEFUL WAY TO EXPERIMENT WITH THE PLATFORM AS IT ALLOWS
-YOU TO QUICKLY BUILD A UI FOR DEMONSTRATION PURPOSES.
-
-GOING FORWARD WE RECOMMEND IMPLEMENTING A STANDALONE WEB SERVER THAT AUTHORISES VIA THE NODE'S RPC INTERFACE. IN THE
-COMING WEEKS, WE'LL WRITE A TUTORIAL ON HOW BEST TO DO THIS.
--->
-
-<!-- Define your frontend here. -->
-
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Template CorDapp</title>
-    <script src="js/template-js.js"></script>
-</head>
-<body>
-<div>Template CorDapp frontend.</div>
-</body>
-</html>
