--- java-source/src/test/java/com/template/contract/TemplateTests.java
@@ -1,3 +1,0 @@
-package com.template.contract;
-
-public class TemplateTests {}
