--- kotlin-source/src/main/kotlin/com/template/contract/TemplateContract.kt
@@ -1,19 +1,0 @@
-package com.template.contract
-
-import net.corda.core.contracts.Contract
-import net.corda.core.contracts.TransactionForContract
-import net.corda.core.crypto.SecureHash
-
-/**
- * Define your contract here.
- */
-open class TemplateContract : Contract {
-    /**
-     * The verify() function of the contract of each of the transaction's input and output states must not throw an
-     * exception for a transaction to be considered valid.
-     */
-    override fun verify(tx: TransactionForContract) {}
-
-    /** A reference to the underlying legal contract template and associated parameters. */
-    override val legalContractReference: SecureHash = SecureHash.sha256("Prose contract.")
-}
