+++ kotlin-source/src/main/kotlin/net/corda/training/Main.kt
