+++ kotlin-source/src/main/kotlin/net/corda/training/flow/CollectSignatureFlow.kt
@@ -1,0 +1,5 @@
+package net.corda.training.flow
+
+/**
+ * Created by rogerwillis on 07/04/2017.
+ */
