+++ kotlin-source/src/main/kotlin/net/corda/training/plugin/IOUPlugin.kt
@@ -1,0 +1,46 @@
+package net.corda.training.plugin
+
+import net.corda.training.api.TemplateApi
+import net.corda.training.flow.TemplateFlow
+import net.corda.training.service.IOUService
+import net.corda.core.messaging.CordaRPCOps
+import net.corda.core.node.CordaPluginRegistry
+import net.corda.core.node.PluginServiceHub
+import net.corda.core.serialization.SerializationCustomization
+import net.corda.training.service.IOUService.Service
+import java.util.function.Function
+
+class TemplatePlugin : CordaPluginRegistry() {
+    /**
+     * A list of classes that expose web APIs.
+     */
+    override val webApis: List<Function<CordaRPCOps, out Any>> = listOf(Function(::TemplateApi))
+
+    /**
+     * A list of flows required for this CorDapp.
+     */
+    override val requiredFlows: Map<String, Set<String>> = mapOf(
+            TemplateFlow.Initiator::class.java.name to setOf()
+    )
+
+    /**
+     * A list of long-lived services to be hosted within the node.
+     */
+    override val servicePlugins: List<Function<PluginServiceHub, out Any>> = listOf(Function(IOUService::Service))
+
+    /**
+     * A list of directories in the resources directory that will be served by Jetty under /web.
+     * The template's web frontend is accessible at /web/template.
+     */
+    override val staticServeDirs: Map<String, String> = mapOf(
+            // This will serve the templateWeb directory in resources to /web/template
+            "template" to javaClass.classLoader.getResource("templateWeb").toExternalForm()
+    )
+
+    /**
+     * Whitelisting the required types for serialisation by the Corda node.
+     */
+    override fun customizeSerialization(custom: SerializationCustomization): Boolean {
+        return true
+    }
+}
