--- kotlin-source/src/test/kotlin/com/template/contract/TemplateTests.kt
@@ -1,3 +1,0 @@
-package com.template.contract
-
-class TemplateTests
