+++ kotlin-source/src/test/kotlin/net/corda/training/contract/IOUIssueTests.kt
@@ -1,0 +1,3 @@
+package net.corda.training.contract
+
+class TemplateTests
