+++ kotlin-source/src/test/kotlin/net/corda/training/contract/IOUTransferTests.kt
@@ -1,0 +1,5 @@
+package net.corda.training.contract
+
+/**
+ * Created by rogerwillis on 07/04/2017.
+ */
