+++ kotlin-source/src/test/kotlin/net/corda/training/flow/IOUIssueFlowTests.kt
@@ -1,0 +1,5 @@
+package net.corda.training.flow
+
+/**
+ * Created by rogerwillis on 07/04/2017.
+ */
