+++ kotlin-source/src/test/kotlin/net/corda/training/state/IOUStateTests.kt
@@ -1,0 +1,5 @@
+package net.corda.training.state
+
+/**
+ * Created by rogerwillis on 06/04/2017.
+ */
